FileSystemTest: move relevant tests from createImageWizard to fileSystem
this commit move relevant tests from createImageWizard to fileSystem test
